feat: update Gallery
