+++ package.json
@@ -6,6 +6,7 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "@types/react-image-gallery": "^1.2.4",
     "axios": "^1.5.0",
     "firebase": "^10.8.1",
     "firebase-tools": "^13.4.1",
@@ -48,6 +49,6 @@
     "@commitlint/config-conventional": "^17.7.0",
     "husky": "^8.0.3",
     "sass": "^1.66.1",
-    "typescript": "^5.4.5"
+    "typescript": "^4.0.0"
   }
 }
